pyproject.toml
[project]
name = "bookly"
version = "0.1.0"
description = ""
authors = [
    {name = "RichardAyalaPH",email = "richard@purrfecthire.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0.0"
dependencies = [
    "fastapi (>=0.119.0,<0.120.0)",
    "uvicorn (>=0.27.0,<0.28.0)",
    "sqlmodel (>=0.0.27,<0.0.28)",
    "sqlalchemy (>=2.0.14,<2.0.36)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.0.0,<5.0.0)",
    "redis[asyncio] (>=5.0.0,<6.0.0)",
    "fastapi-mail (>=1.5.8,<2.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "bookly", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
